Extract tab site-permission handling and teardown
